button to change song using onclick(this.id)
successfully change audio , but the text name of song etc , image (infuture) is not changed at the same time.
if play top 50 song , cookie will save current time and song but will it continue playing the rest of playlist or will hav error? the for loop is for current playlist to run the song , or when onclick from top 50 add new song to the array to play?